refactor duplicate code to base class; rename params.json
