+++ taxcalc/behavior.py
@@ -3,6 +3,7 @@ import json
 import os
 import numpy as np
 from .parameters import Parameters
+from .parameters_base import ParametersBase
 from .utils import expand_array
 
 
@@ -77,10 +78,10 @@ def behavior(calc_x, calc_y, update_income=update_income):
     return calc_y_behavior
 
 
-class Behavior(object):
+class Behavior(ParametersBase):
 
     JSON_START_YEAR = Parameters.JSON_START_YEAR
-    BEHAVIOR_FILENAME = 'behavior.json'
+    DEFAULTS_FILENAME = 'behavior.json'
     DEFAULT_NUM_YEARS = Parameters.DEFAULT_NUM_YEARS
 
     def __init__(self, behavior_dict=None,
@@ -92,73 +93,14 @@ class Behavior(object):
                 raise ValueError('behavior_dict is not a dictionary')
             self._vals = behavior_dict
         else:  # if None, read current-law parameters
-            self._vals = self._behavior_dict_from_json_file()
+            self._vals = self._params_dict_from_json_file()
 
-        if num_years < 1:
-            raise ValueError('num_years < 1')
-
-        self._current_year = start_year
-        self._start_year = start_year
-        self._num_years = num_years
-        self._end_year = start_year + num_years - 1
-        self.set_default_vals()
-
-    def set_default_vals(self):
-        # extend current-law parameter values into future with _inflation_rates
-        for name, data in self._vals.items():
-            values = data['value']
-            setattr(self, name,
-                    expand_array(values, inflate=False,
-                                 inflation_rates=None,
-                                 num_years=self._num_years))
-        self.set_year(self._start_year)
-
-    @property
-    def num_years(self):
-        return self._num_years
-
-    @property
-    def current_year(self):
-        return self._current_year
-
-    @property
-    def end_year(self):
-        return self._end_year
-
-    @property
-    def start_year(self):
-        return self._start_year
-
-    @staticmethod
-    def _behavior_dict_from_json_file():
-        """
-        Read params.json file and return complete params dictionary.
-
-        Parameters
-        ----------
-        nothing: void
-
-        Returns
-        -------
-        params: dictionary
-            containing complete contents of params.json file.
-        """
-        behavior_path = os.path.join(os.path.abspath(
-                                     os.path.dirname(__file__)),
-                                     Behavior.BEHAVIOR_FILENAME)
-        if os.path.exists(behavior_path):
-            with open(behavior_path) as pfile:
-                behavior = json.load(pfile)
-        else:
-            from pkg_resources import resource_stream, Requirement
-            path_in_egg = os.path.join('taxcalc', Behavior.BEHAVIOR_FILENAME)
-            buf = resource_stream(Requirement.parse('taxcalc'), path_in_egg)
-            as_bytes = buf.read()
-            as_string = as_bytes.decode("utf-8")
-            behavior = json.loads(as_string)
-        return behavior
+        self.initialize(start_year, num_years)
 
     def update_behavior(self, reform):
+        '''Update behavior for a reform, a dictionary
+        consisting of year: modification dictionaries. For example:
+        {2014: {'_BE_inc': [0.4, 0.3]}}'''
         self.set_default_vals()
         if self.current_year != self.start_year:
             self.set_year(self.start_year)
@@ -166,52 +108,4 @@ class Behavior(object):
         for year in reform:
             if year != self.start_year:
                 self.set_year(year)
-            num_years_to_expand = (self.start_year + self.num_years) - year
-            for name, values in reform[year].items():
-                # determine inflation indexing status of parameter with name
-                cval = getattr(self, name, None)
-                if cval is None:
-                    # it is a tax law parameter not behavior
-                    continue
-                nval = expand_array(values,
-                                    inflate=False,
-                                    inflation_rates=None,
-                                    num_years=num_years_to_expand)
-
-                cval[(self.current_year - self.start_year):] = nval
-        self.set_year(self.start_year)
-
-    def set_year(self, year):
-        """
-        Set behavior parameters to values for specified calendar year.
-
-        Parameters
-        ----------
-        year: int
-            calendar year for which to current_year and parameter values
-
-        Raises
-        ------
-        ValueError:
-            if year is not in [start_year, end_year] range.
-
-        Returns
-        -------
-        nothing: void
-
-        Notes
-        -----
-        To increment the current year, use the following statement::
-
-            behavior.set_year(behavior.current_year + 1)
-
-        where behavior is a policy Behavior object.
-        """
-        if year < self.start_year or year > self.end_year:
-            msg = 'year passed to set_year() must be in [{},{}] range.'
-            raise ValueError(msg.format(self.start_year, self.end_year))
-        self._current_year = year
-        year_zero_indexed = year - self._start_year
-        for name in self._vals:
-            arr = getattr(self, name)
-            setattr(self, name[1:], arr[year_zero_indexed])
+            self._update({year: reform[year]})
